Rename src folder to lib
