+++ types/three-gif-loader/lib/gif-texture.d.ts
